# backend/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Vote Électronique",
        "description": "API pour le système de vote électronique",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/api/elections/{election_id}/candidats": {
            "post": {
                "tags": [
                    "Candidats"
                ],
                "summary": "Ajouter un candidat à une élection",
                "operationId": "4c8f5ae23c0a2ab7b0eaabeea86601c1",
                "parameters": [
                    {
                        "name": "election_id",
                        "in": "path",
                        "description": "ID de l'élection",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "programme": {
                                        "type": "string",
                                        "example": "Mon programme électoral"
                                    },
                                    "photo_path": {
                                        "type": "string",
                                        "example": "candidats/photo.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidat ajouté avec succès"
                    },
                    "404": {
                        "description": "Élection non trouvée"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidats/{id}": {
            "put": {
                "tags": [
                    "Candidats"
                ],
                "summary": "Modifier un candidat",
                "operationId": "7e4a8fddbea4cdc6200e832bc0ba81ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du candidat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "programme": {
                                        "type": "string",
                                        "example": "Mon programme mis à jour"
                                    },
                                    "photo": {
                                        "description": "Photo du candidat",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidat mis à jour avec succès"
                    },
                    "403": {
                        "description": "Élection déjà ouverte ou clôturée"
                    },
                    "404": {
                        "description": "Candidat non trouvé"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Candidats"
                ],
                "summary": "Supprimer un candidat",
                "operationId": "e53e50c31438415655e635357aa40567",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du candidat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidat supprimé avec succès"
                    },
                    "403": {
                        "description": "Élection déjà ouverte ou clôturée"
                    },
                    "404": {
                        "description": "Candidat non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/upload/photo": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "Uploader une photo de candidat",
                "operationId": "cfe8e7b31a665066d0cbcaa1354013e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "description": "Photo du candidat (jpeg, jpg, png - max 2Mo)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo uploadée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "photo_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/storage/candidats/photo.jpg"
                                        },
                                        "photo_path": {
                                            "type": "string",
                                            "example": "candidats/photo.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fichier invalide ou manquant"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}